requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
cycler==0.10.0
emoji==1.4.1
googletrans==4.0.0rc1
greenlet==1.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kiwisolver==1.3.1
kombu==5.1.0
loader==2017.9.11
lxml==4.6.3
matplotlib==3.4.2
multidict==5.1.0
numpy==1.21.1
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
pywin32==301
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
future==0.18.2
pydub==0.25.1
SpeechRecognition==3.8.1
yarl==1.6.3


fuzzywuzzy~=0.18.0